+++ stac-api/src/filter.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 
 /// The language of the filter expression.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[serde(tag = "filter-lang", content = "filter")]
 #[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub enum Filter {